Dynamic Images For Desktop and Mobile
The images will now change when the width of the device viewing the page is less than 600px wide. Anything greater than 600px wide will display the desktop image.

I've also added the off button to the modal. 

The on button still needs to be added and the javascript to change the title's text.