Modal Created
The modal has been completed.

a) The modal will be turned on through a button within the modal labeled "ON". When the modal is on, it can be turned off through the "OFF" button outside of the modal.
b) The modal has an image and a title
c) On mobile, the image is replaced with CSS. Technically, the image for mobile and desktop are loaded at the same time when the site is visited. With the display property, I show and hide images based on the width of the device the visitor is using.
d) On mobile, the title is changed with Javascript. Just in case, the title is also changed with Javascript on page load.
e) All source code and changes in this repository.